Reviewed and updated page 04.rst
Global comment: Please comply with the guidelines provided in the Intel-Names Database for the usage and name of the Universal Wellpad Controller. See, https://prod-namesdb.intel.com/ 
The abbreviation for Universal Wellpad Controller should be expanded in the menu options of the software. After this change is implemented we can align the documentation as well. FYR see lines 52 to 61.

Also, in the menu options, micro-services should be rectified to microservices. Hyphen is not required for 'microservices'.