Reviewed and updated page_06.rst
Suggestion: The following terms 'master' and 'slave' need to be replaced with more inclusive words such as primary or main and secondary. If this does not impact any code or labels then we can change this term to be more inclusive.

Hi @nagdeepgk, Please confirm if we can proceed with this change.